LayoutTests/http/tests/contentextensions/loading/main-resource-redirect-blocked-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didFailProvisionalLoadWithError
main frame - (kWKErrorCodeCannotShowURL)
